+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/eyecandy/EyeCandyScriptContext.java
@@ -4,6 +4,7 @@ import com.lx862.jcm.mod.scripting.mtr.MTRScriptContext;
 import com.lx862.jcm.mod.scripting.mtr.render.ModelDrawCall;
 import com.lx862.jcm.mod.scripting.mtr.util.ScriptedModel;
 import com.lx862.mtrscripting.util.Matrices;
+import com.lx862.mtrscripting.util.Vector3dWrapper;
 import org.mtr.mapping.holder.Identifier;
 
 import java.util.ArrayList;
@@ -24,6 +25,6 @@ public class EyeCandyScriptContext extends MTRScriptContext {
     }
 
     public void playSound(Identifier id, float volume, float pitch) {
-        soundManager().playSound(id, blockEntity.pos(), volume, pitch);
+        soundManager().playSound(id, Vector3dWrapper.ZERO, volume, pitch);
     }
 }