+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/render/ModelDrawCall.java
@@ -2,6 +2,7 @@ package com.lx862.jcm.mod.scripting.mtr.render;
 
 import com.lx862.mtrscripting.api.ScriptResultCall;
 import com.lx862.jcm.mod.scripting.mtr.util.ScriptedModel;
+import com.lx862.mtrscripting.util.Vector3dWrapper;
 import org.mtr.mapping.holder.Direction;
 import org.mtr.mapping.holder.World;
 import org.mtr.mapping.mapper.GraphicsHolder;
@@ -17,9 +18,10 @@ public class ModelDrawCall extends ScriptResultCall {
     }
 
     @Override
-    public void run(World world, GraphicsHolder graphicsHolder, StoredMatrixTransformations storedMatrixTransformations, Direction facing, int light) {
+    public void run(World world, Vector3dWrapper basePos, GraphicsHolder graphicsHolder, StoredMatrixTransformations storedMatrixTransformations, Direction facing, int light) {
         if(this.storedMatrixTransformations != null) {
             storedMatrixTransformations.add(this.storedMatrixTransformations);
+            storedMatrixTransformations.add(gh -> gh.translate(basePos.x(), basePos.y(), basePos.z()));
         }
 
         this.model.draw(storedMatrixTransformations, light);