+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/sound/PositionedSoundCall.java
@@ -1,5 +1,6 @@
 package com.lx862.jcm.mod.scripting.mtr.sound;
 
+import com.lx862.mtrscripting.util.Vector3dWrapper;
 import org.mtr.mapping.holder.*;
 import org.mtr.mapping.mapper.GraphicsHolder;
 import org.mtr.mapping.mapper.SoundHelper;
@@ -18,7 +19,9 @@ public class PositionedSoundCall extends SoundCall {
     }
 
     @Override
-    public void run(World world, GraphicsHolder graphicsHolder, StoredMatrixTransformations storedMatrixTransformations, Direction facing, int light) {
-        world.playSound(x, y, z, soundEvent, SoundCategory.valueOf(soundCategory), volume, pitch, false);
+    public void run(World world, Vector3dWrapper basePos, GraphicsHolder graphicsHolder, StoredMatrixTransformations storedMatrixTransformations, Direction facing, int light) {
+        Vector3dWrapper finalPos = basePos.copy();
+        finalPos.add((float)x, (float)y, (float)z);
+        world.playSound(finalPos.x(), finalPos.y(), finalPos.z(), soundEvent, SoundCategory.valueOf(soundCategory), volume, pitch, false);
     }
 }