+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/sound/ScriptSoundManager.java
@@ -18,9 +18,9 @@ public class ScriptSoundManager {
         this.soundCalls = new ArrayList<>();
     }
 
-    public void updateSoundCalls(ScriptSoundManager other) {
-        soundCalls.clear();
-        soundCalls.addAll(other.soundCalls);
+    private ScriptSoundManager(List<ScriptResultCall> soundCalls) {
+        this.soundCalls = new ArrayList<>();
+        this.soundCalls.addAll(soundCalls);
     }
 
     public void playLocalSound(Identifier id, float volume, float pitch) {
@@ -31,10 +31,14 @@ public class ScriptSoundManager {
         this.soundCalls.add(new PositionedSoundCall(id, pos.x(), pos.y(), pos.z(), volume, pitch));
     }
 
-    public void invoke(World world, GraphicsHolder graphicsHolder, StoredMatrixTransformations storedMatrixTransformations, Direction facing, int light) {
+    public void invoke(World world, Vector3dWrapper basePos, GraphicsHolder graphicsHolder, StoredMatrixTransformations storedMatrixTransformations, Direction facing, int light) {
         for(ScriptResultCall soundCall : new ArrayList<>(soundCalls)) {
-            soundCall.run(world, graphicsHolder, storedMatrixTransformations, facing, light);
+            soundCall.run(world, basePos, graphicsHolder, storedMatrixTransformations, facing, light);
         }
+    }
+
+    public ScriptSoundManager copy() {
+        return new ScriptSoundManager(this.soundCalls);
     }
 
     public void reset() {