+++ fabric/src/main/java/com/lx862/jcm/mod/scripting/mtr/vehicle/VehicleScriptInstance.java
@@ -1,38 +1,34 @@
 package com.lx862.jcm.mod.scripting.mtr.vehicle;
 
+import com.lx862.jcm.mod.scripting.mtr.render.ScriptRenderManager;
 import com.lx862.jcm.mod.scripting.mtr.sound.ScriptSoundManager;
-import com.lx862.jcm.mod.scripting.mtr.sound.SoundCall;
 import com.lx862.mtrscripting.core.ParsedScript;
 import com.lx862.mtrscripting.core.ScriptInstance;
 import org.mtr.mod.client.MinecraftClientData;
 import org.mtr.mod.data.VehicleExtension;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class VehicleScriptInstance extends ScriptInstance<VehicleWrapper> {
     private final VehicleExtension vehicleExtension;
-    public final List<VehicleModelDrawCall> carModelDrawCalls;
-    private final ScriptSoundManager scriptSoundManager;
+    public final ScriptRenderManager[] renderManagers;
+    public final ScriptSoundManager[] soundManagers;
 
     public VehicleScriptInstance(VehicleScriptContext context, VehicleExtension vehicleExtension, ParsedScript script) {
         super(context, script);
         this.vehicleExtension = vehicleExtension;
-        this.carModelDrawCalls = new ArrayList<>();
-        this.scriptSoundManager = new ScriptSoundManager();
+        this.renderManagers = new ScriptRenderManager[vehicleExtension.vehicleExtraData.immutableVehicleCars.size()];
+        this.soundManagers = new ScriptSoundManager[vehicleExtension.vehicleExtraData.immutableVehicleCars.size()];
     }
 
-    public void setCarModelDrawCalls(List<VehicleModelDrawCall> calls) {
-        this.carModelDrawCalls.clear();
-        this.carModelDrawCalls.addAll(calls);
+    public void saveRenderCalls(ScriptRenderManager[] renderManagers) {
+        for(int i = 0; i < renderManagers.length; i++) {
+            this.renderManagers[i] = renderManagers[i].copy();
+        }
     }
 
-    public void updateSound(ScriptSoundManager soundManager) {
-        scriptSoundManager.updateSoundCalls(soundManager);
-    }
-
-    public ScriptSoundManager getSoundManager() {
-        return this.scriptSoundManager;
+    public void saveSoundCalls(ScriptSoundManager[] soundManagers) {
+        for(int i = 0; i < soundManagers.length; i++) {
+            this.soundManagers[i] = soundManagers[i].copy();
+        }
     }
 
     public boolean shouldInvalidate() {