+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - - **Eyecandy Scripting**
 - - - Adapt RenderManager, can be accessed by `ctx.renderManager()`, see above.
 - - - Adapt SoundManager, can be accessed by `ctx.soundManager()`, see above.
-- - - Add `EyecandyBlockEntity.pos()` and `EyecandyBlockEntity.blockPos()` to determine where the eyecandy block is placed.
+- - - Add `EyecandyBlockEntity.pos()`, `EyecandyBlockEntity.blockPos()` and `EyecandyBlockEntity.offsetPos()` to determine where the eyecandy block is placed.
 - - **PIDS Scripting**
 - - - Adapt RenderManager, can be accessed by `ctx.renderManager()`, see above.
 - - - - This grants PIDS scripts the ability to render 3D OBJ model. It is advised to use this in conjunction with PIDS Projector, which is an invisible block.